dev mode

+++ test/new-e2e/tests/agent-metric-logs/log-agent/utils/file_tailing_utils.go
@@ -31,6 +31,7 @@ type LogTestHelper interface {
 	CheckLogs(service, content string, expectLogs bool)
 	CleanUp()
 	EventuallyWithT(assertFunction func(*assert.CollectT), waitFor, tick time.Duration, option ...interface{}) bool
+	DevMode() bool
 }
 
 // AppendLog appen log with 'content', which is then repeated 'reccurrence' times and verifies log contents.
@@ -136,27 +137,27 @@ func CleanUp(s LogTestHelper) {
 	t := s.T()
 	var checkCmd string
 
-	// if s.DevMode == true {
-	osType := s.Env().VM.GetOSType()
+	if s.DevMode == func()true {
+		osType := s.Env().VM.GetOSType()
 
-	switch osType {
-	default: // default is linux
-		s.Env().VM.Execute("sudo rm -f /var/log/hello-world.log")
-		s.Env().VM.Execute("sudo rm -f /var/log/hello-world.log.old")
-		checkCmd = "ls /var/log/hello-world.log /var/log/hello-world.log.old 2>/dev/null || echo 'Files do not exist'"
-	case componentos.WindowsType:
-		s.Env().VM.Execute("if (Test-Path C:\\logs\\hello-world.log) { Remove-Item -Path C:\\logs\\hello-world.log -Force }")
-		s.Env().VM.Execute("if (Test-Path C:\\logs\\hello-world.log.old) { Remove-Item -Path C:\\logs\\hello-world.log.old -Force }")
-		checkCmd = "if (Test-Path C:\\logs\\hello-world.log) { Get-ChildItem -Path C:\\logs\\hello-world.log } elseif (Test-Path C:\\logs\\hello-world.log.old) { Get-ChildItem -Path C:\\logs\\hello-world.log.old } else { Write-Output 'Files do not exist' }"
-	}
-
-	s.EventuallyWithT(func(c *assert.CollectT) {
-		output, err := s.Env().VM.ExecuteWithError(checkCmd)
-		if assert.NoErrorf(c, err, "Having issue cleaning up log files, retrying... %s", output) {
-			t.Log("Successfully cleaned up log files.")
+		switch osType {
+		default: // default is linux
+			s.Env().VM.Execute("sudo rm -f /var/log/hello-world.log")
+			s.Env().VM.Execute("sudo rm -f /var/log/hello-world.log.old")
+			checkCmd = "ls /var/log/hello-world.log /var/log/hello-world.log.old 2>/dev/null || echo 'Files do not exist'"
+		case componentos.WindowsType:
+			s.Env().VM.Execute("if (Test-Path C:\\logs\\hello-world.log) { Remove-Item -Path C:\\logs\\hello-world.log -Force }")
+			s.Env().VM.Execute("if (Test-Path C:\\logs\\hello-world.log.old) { Remove-Item -Path C:\\logs\\hello-world.log.old -Force }")
+			checkCmd = "if (Test-Path C:\\logs\\hello-world.log) { Get-ChildItem -Path C:\\logs\\hello-world.log } elseif (Test-Path C:\\logs\\hello-world.log.old) { Get-ChildItem -Path C:\\logs\\hello-world.log.old } else { Write-Output 'Files do not exist' }"
 		}
-	}, 1*time.Minute, 10*time.Second)
-	// }
+
+		s.EventuallyWithT(func(c *assert.CollectT) {
+			output, err := s.Env().VM.ExecuteWithError(checkCmd)
+			if assert.NoErrorf(c, err, "Having issue cleaning up log files, retrying... %s", output) {
+				t.Log("Successfully cleaned up log files.")
+			}
+		}, 1*time.Minute, 10*time.Second)
+	}
 
 	s.EventuallyWithT(func(c *assert.CollectT) {
 		err := s.Env().Fakeintake.FlushServerAndResetAggregators()
